+++ docs/src/env.d.ts
@@ -1,3 +1,9 @@
 /// <reference path="../.astro/types.d.ts" />
 /// <reference types="astro/client" />
 /// <reference types="@astrojs/starlight/virtual" />
+
+declare namespace App {
+  interface Locals {
+    starlight: import('@astrojs/starlight/props').Props;
+  }
+}