+++ packages/aik-mod/CHANGELOG.md
@@ -7,6 +7,13 @@
 - Updated dependencies [aa73961]
   - @inox-tools/inline-mod@1.5.3
 
+## 0.8.5-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+  - @inox-tools/inline-mod@2.0.0-beta.0
+
 ## 0.8.4
 
 ### Patch Changes