+++ packages/aik-route-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inox-tools/aik-route-config
 
+## 0.7.5-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+- Updated dependencies [b4843b9]
+  - @inox-tools/utils@0.3.0-beta.0
+
 ## 0.7.4
 
 ### Patch Changes