+++ packages/inline-mod/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inox-tools/inline-mod
 
+## 2.0.0-beta.0
+
+### Major Changes
+
+- b4843b9: Bump support to Vite 6
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+- Updated dependencies [b4843b9]
+  - @inox-tools/utils@0.3.0-beta.0
+
 ## 1.5.3
 
 ### Patch Changes