+++ packages/portal-gun/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @inox-tools/portal-gun
 
+## 1.0.0-beta.0
+
+### Major Changes
+
+- b4843b9: Bump support to Astro 5
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+- Updated dependencies [b4843b9]
+  - @inox-tools/utils@0.3.0-beta.0
+  - @inox-tools/runtime-logger@0.3.5-beta.0
+
 ## 0.1.1
 
 ### Patch Changes