+++ packages/star-warp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inox-tools/star-warp
 
+## 0.1.6-beta.0
+
+### Patch Changes
+
+- Updated dependencies [b4843b9]
+- Updated dependencies [b4843b9]
+  - @inox-tools/utils@0.3.0-beta.0
+
 ## 0.1.5
 
 ### Patch Changes